alter tables to use set type columns
the st type helps guarantee correct values and assists in insertion of data, however it does not use the pipe character to delimit values, instead it uses a comma, which means that we have to adjust our code to split or explode on a comma instead of on the pipe. This goes for liturgical colors and liturgical commons. we also starting adding DIOCESIDIROMA data